Move package metadata functionality from package to mirror
Clean up the public API of Package class
